1.2.91 - Inicialização dos serviços de CharServer, solicitação de acesso com LoginServer. - LoginSeed corrigido, first e second devem ser do tipo int ao invés de long. - CT_NONE adicionado para ClientType não definido (será usado mais a frente). - Envio de pacote "ping" para com CharServer, ServiceLoginClient.pingCharRequest.

+++ JRagnarok Emulador/src/org/diverproject/jragnarok/server/login/structures/LoginSeed.java
@@ -6,10 +6,10 @@ import org.diverproject.util.ObjectDescription;
 
 public class LoginSeed
 {
-	private long first;
-	private long second;
+	private int first;
+	private int second;
 
-	public long getFirst()
+	public int getFirst()
 	{
 		return first;
 	}
@@ -19,7 +19,7 @@ public class LoginSeed
 		first = random() + 1;
 	}
 
-	public long getSecond()
+	public int getSecond()
 	{
 		return second;
 	}
